requirements.txt
flask
werkzeug
python-docx
pdfkit
PyPDF2
